layered: fixed gwt compilation issue #179
 Type mismatch: cannot convert from Integer to double

+++ plugins/org.eclipse.elk.alg.layered/src/org/eclipse/elk/alg/layered/graph/LGraphUtil.java
@@ -850,12 +850,12 @@ public final class LGraphUtil {
                 switch (finalExternalPortSide) {
                 case NORTH:
                 case EAST:
-                    informationAboutIt = propertyHolder.getProperty(LayeredOptions.PORT_INDEX);
+                    informationAboutIt = propertyHolder.getProperty(LayeredOptions.PORT_INDEX).doubleValue();
                     break;
                     
                 case SOUTH:
                 case WEST:
-                    informationAboutIt = -1 * propertyHolder.getProperty(LayeredOptions.PORT_INDEX);
+                    informationAboutIt = -1 * propertyHolder.getProperty(LayeredOptions.PORT_INDEX).doubleValue();
                     break;
                 }
